Pull up method to TestSuiteDescriptorLoader level from JsonBased one's.
